fix(gql): rename graphql app to gql to avoid naming conflict
BREAKING: Rename GraphQL Django app from 'graphql' to 'gql' to prevent
import collision with Python's graphql-core library.

Issue:
When starting Docker containers, Django failed to import strawberry_django
because the 'graphql' app name conflicted with the 'graphql' package from
graphql-core (Strawberry's dependency). This caused import errors during
Django setup.

Root cause:
Python's module resolution found our 'graphql' Django app instead of the
'graphql' package from graphql-core, breaking Strawberry's imports.

Solution:
Rename app to 'gql' (short, clear, no conflicts):
- graphql/ → gql/ directory
- GraphqlConfig → GqlConfig class
- Update all imports and references
- Update documentation

Changes:
- Rename directory: graphql/ → gql/
- gql/apps.py: GraphqlConfig → GqlConfig, name='gql'
- gql/__init__.py: Update default_app_config path
- datahub/settings.py: graphql.apps.GraphqlConfig → gql.apps.GqlConfig
- datahub/urls.py: from graphql.schema → from gql.schema
- demo_graphql.py: Update import path
- gql/README.md: Update all references to app name and paths
- gql/DEMO_QUERIES.md: No changes needed (queries unchanged)

Benefits:
- Resolves Docker startup error
- No naming conflicts with Python packages
- Still clear and concise (gql is common abbreviation for GraphQL)
- Consistent with commit scope naming convention

Fixes Docker import error that prevented containers from starting.
