+++ gql/DEMO_QUERIES.md
