+++ gql/apps.py
@@ -4,11 +4,11 @@ Django app configuration for GraphQL API.
 from django.apps import AppConfig
 
 
-class GraphqlConfig(AppConfig):
+class GqlConfig(AppConfig):
     """Configuration for the GraphQL app."""
     
     default_auto_field = 'django.db.models.BigAutoField'
-    name = 'graphql'
+    name = 'gql'
     verbose_name = 'GraphQL API'
     
     def ready(self):
